+++ src/main/java/game/Game.java
@@ -10,10 +10,9 @@ import game.event.JoinListener;
 import game.event.LeaveListener;
 import game.event.TeamScoreListener;
 import game.event.TeamSelectListener;
-import game.hockey.HockeyGame;
 import game.states.GameStateContext;
 
-public interface Game extends GameStateContext, HockeyGame {
+public interface Game extends GameStateContext {
 	
 	void tick();
