--- src/main/java/game/hockey/HockeyGame.java
@@ -1,11 +1,0 @@
-package game.hockey;
-
-public interface HockeyGame {
-
-	void addGoal(Goal goal);
-	
-//	void onTeamScored(String team, int points);
-		
-	Goal findGoalOfTeam(String team);
-	
-}
