+++ src/main/java/usecases/hockey/savegame/validation/AllGoalsShouldBeSetRule.java
@@ -2,6 +2,7 @@ package usecases.hockey.savegame.validation;
 
 import entities.Team;
 import game.Game;
+import game.hockey.HockeyGameCycle;
 import usecases.hockey.savegame.SaveGame.SaveGameResponse;
 import usecases.hockey.savegame.validation.SaveGameValidator.SaveGameValidationRule;
 
@@ -22,8 +23,9 @@ public class AllGoalsShouldBeSetRule implements SaveGameValidationRule {
 	}
 	
 	private boolean notAllGoalsSet() {
+		HockeyGameCycle hockey = (HockeyGameCycle) game.getGameCycle();
 		for (Team team : game.getTeams().findAllTeams()) {
-			if (game.findGoalOfTeam(team.getName()) == null)
+			if (hockey.findGoalOfTeam(team.getName()) == null)
 				return true;
 		}
 		return false;
